+++ lib/getMetrikaInstance.ts
@@ -1,8 +1,8 @@
-import { Config, YaMetrika } from "./types"
+import type { Config, YaMetrika } from "./types"
 
-export function getMetrikaInstance (config: Config & Record<string, any>): YaMetrika {
+export function getMetrikaInstance (config: Config | Record<string, any>): YaMetrika {
     const instance = new window.Ya.Metrika2(config)
-    setMetrikaInstance(instance, config)
+    setMetrikaInstance(instance, config as Config)
     return instance
 }
